+++ app/src/Constants/MenuItems.tsx
@@ -1,0 +1,8 @@
+const menuItems = [
+  { name: 'home', route: '/' },
+  { name: 'Schedule', route: '/schedule' },
+  { name: 'Tournaments', route: '/tournaments' },
+  { name: 'Standings', route: '/standings' },
+];
+
+export default menuItems;
